Wrap ttLib cmap code that constructs format 12/13 tables.
Currently some code in the _c_m_a_p file of the copy of ttLib under gae_server
is commented out / changed to avoid coalescing cmap runs.  In order to use
the upstream ttLib, the quick thing to do is handle this like the other ttLib
patches.

The only call that relies on this in the build_time is in cleanup.py.

This does not revert the original corresponding patch to _c_m_a_p
since I do not know if there's any runtime code that relies on it.  That
patch also changed an additional function, which this change does not
reimplement since it is not called from the build time code.